a bunch of updates

+++ main/plugins/p002_dht.h
@@ -7,12 +7,12 @@
 
 class DHTPlugin: public Plugin {
     private:
-        float temperature = 0;
-        float humidity = 0;
+        double temperature = 0;
+        double humidity = 0;
         Type temperature_t = Type::decimal;
         Type humidity_t = Type::decimal;
 
-        float temp[2];
+        double temp[2];
         TaskHandle_t task_h;
     public:
         DEFINE_PPLUGIN(DHTPlugin, 2);
